fix bug

+++ PlagiarismDetection/Services/IVectorStore.cs
@@ -4,6 +4,6 @@
     {
         Task UpsertAsync(string id, float[] vector, object metadata);
 
-        Task<IEnumerable<(string Id, float Score, dynamic Metadata)>> QueryAsync(float[] vector, int topK = 10);
+        Task<IEnumerable<(string Id, float Score, dynamic Metadata)>> QueryAsync(float[] vector, ulong topK = 10);
     }
 }
